refactor: s/crash-reporter/reporter/
